Handle unique user registration
There is an error where when a person double clicks the user
registration form, it will send two requests to create an account.
Since at the moment, both requests are valid, the uniqueness will pass.

Here I add a check of the atomic transaction to check if there
exists a user with the username and password, and if there
exists, return it with a flag saying that the account was not created.

I am using that flag to not send again the user activation email.